backend/ml/requirements.txt
flask==2.2.5
pymongo==4.6.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
geopy==2.4.1
